ml-consumer/requirements.txt
confluent-kafka==2.3.0
Flask
uvicorn
requests==2.31.0
python-dotenv==1.0.1
scikit-learn==1.3.2
numpy==1.26.4
joblib==1.3.2